clean up css and js
